Refactor `forc_pkg::BuildOptions` into its parts (#3166)
This is in anticipation of introducing a new `forc-test` library crate
that will contain the logic for building and executing tests. It will
expose a similar `TestOpts` type that shares some of these parts, but
not all.

Also splits up the forc `BuildCommand` for easier sharing between sets
of clap arguments (using the `clap` `flatten` attribute) that require
building a project (e.g. build, test).

Also adds two missing args to the `RunCommand` (`release` and
`build-profile`).

Co-authored-by: Mohammad Fawaz <mohammadfawaz89@gmail.com>