Remove some uses of `ptr_to_int` and `int_to_ptr`, using `ptr` instead (#7297)
This is in preparation for the argument mutability analysis. This PR
also has some skeleton code for this analysis, but the actual analysis
is still a `todo!()`.

The analysis will also have better precision if we use fewer asm blocks,
but I'll get to that after working on the analysis itself.